chore(admin): Split modules
